+++ src/components/DatatableV2/hooks/useDatatable.ts
@@ -12,7 +12,7 @@ import {
   getSortedRowModel,
   useReactTable,
 } from '@tanstack/react-table';
-import { useMemo, useState } from 'react';
+import { useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react';
 
 import {
   getAllLeafColumnDefs,
@@ -36,6 +36,7 @@ export const useDatatable = <D>(
 ): DatatableInstance<D> => {
   const tableOptions = useOptions<D>(options);
   const displayColumns = useDisplayColumns<D>(tableOptions);
+  const tableRef = useRef<HTMLTableElement>();
 
   const columnDefs = useMemo(
     () =>
@@ -82,6 +83,20 @@ export const useDatatable = <D>(
   const [columnSizing, setColumnSizing] = useState<ColumnSizingState>(
     initialState?.columnSizing ?? {},
   );
+  const [width, setWidth] = useState(0);
+
+  const onResize = () => {
+    setWidth(tableRef.current.getBoundingClientRect().width);
+  };
+  useEffect(() => {
+    window.addEventListener('resize', onResize);
+    return () => {
+      window.removeEventListener('resize', onResize);
+    };
+  }, []);
+  useLayoutEffect(() => {
+    onResize();
+  }, []);
 
   const debouncedSetColumnSizing = useDebounce(setColumnSizing);
 
@@ -124,6 +139,10 @@ export const useDatatable = <D>(
       showColumnSettings,
       columnSizing,
       ...tableOptions.state,
+      // I know what I'm doing here
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
+      width,
     },
     getCoreRowModel: getCoreRowModel(),
     getPaginationRowModel: tableOptions.enablePagination
@@ -136,6 +155,10 @@ export const useDatatable = <D>(
       ? getExpandedRowModel()
       : undefined,
   }) as unknown as DatatableInstance<D>;
+
+  table.refs = {
+    tableRef,
+  };
 
   table.setShowColumnSettings =
     tableOptions.onShowColumnSettings ?? setShowColumnSettings;