+++ src/components/DatatableV2/table/Table.tsx
@@ -14,6 +14,7 @@ const Table = <D,>({ table }: { table: DatatableInstance<D> }) => {
     getFlatHeaders,
     getState,
     options: { columns },
+    refs: { tableRef },
   } = table;
   const {
     columnSizing,
@@ -39,9 +40,15 @@ const Table = <D,>({ table }: { table: DatatableInstance<D> }) => {
   }, [columns, columnSizing, columnSizingInfo, columnVisibility]);
 
   return (
-    <TableRoot>
+    <TableRoot tabIndex={0}>
       {showProgress && <ProgressBar isTop />}
-      <table className="ds-table" style={columnSizeVars}>
+      <table
+        ref={(ref) => {
+          tableRef.current = ref;
+        }}
+        className="ds-table"
+        style={columnSizeVars}
+      >
         <Header table={table} />
         <Body table={table} />
       </table>