+++ src/components/DatatableV2/toolbar/Selection.tsx
@@ -7,10 +7,11 @@ import { abbreviateNumber, pxToRem } from '../../../utils';
 import { Inline, Padbox } from '../../layout';
 import { Button } from '../../Button';
 import SelectButton from '../buttons/SelectButton';
+import { useHasHozintalScroll } from '../hooks/useHasHozintalScroll';
 
-const SelectionRoot = styled(Padbox)`
+const SelectionRoot = styled(Padbox)<{ $hasHorizontalScroll: boolean }>`
   position: absolute;
-  top: 0;
+  top: ${({ $hasHorizontalScroll }) => ($hasHorizontalScroll ? '1.5rem' : 0)};
   left: 0;
   right: 0;
   z-index: 1;
@@ -26,6 +27,8 @@ const Selection = <D,>({ table }: { table: DatatableInstance<D> }) => {
     toggleAllRowsSelected,
   } = table;
 
+  const hasHorizontalScroll = useHasHozintalScroll(table);
+
   const selectedRows = getSelectedRowModel().rows;
   const selectedRowsCount = selectedRows.length;
   const totalRowCount = rowCount ?? getPrePaginationRowModel().rows.length;
@@ -36,6 +39,7 @@ const Selection = <D,>({ table }: { table: DatatableInstance<D> }) => {
 
   return (
     <SelectionRoot
+      $hasHorizontalScroll={hasHorizontalScroll}
       className="ds-table-selection-toolbar"
       paddingSize="mdPlus"
       paddingType="squish"